fix: endpoint 수정
