+++ src/test/java/doore/restdocs/docs/DocumentDocsTest.java
@@ -1,13 +1,14 @@
 package doore.restdocs.docs;
 
-import static doore.document.domain.DocumentAccessType.*;
+import static doore.document.domain.DocumentAccessType.ALL;
+import static doore.document.domain.DocumentAccessType.TEAM;
 import static doore.document.domain.DocumentType.IMAGE;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.delete;
-import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.multipart;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.get;
+import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.multipart;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.put;
 import static org.springframework.restdocs.payload.PayloadDocumentation.requestFields;
 import static org.springframework.restdocs.payload.PayloadDocumentation.requestPartFields;
@@ -30,7 +31,6 @@ import java.io.FileInputStream;
 import java.time.LocalDate;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
-
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.data.domain.Page;
@@ -221,5 +221,40 @@ public class DocumentDocsTest extends RestDocsTest {
                                 parameterWithName("documentId").description("학습자료 id")
                         )
                 ));
+    }
+
+    @Test
+    @DisplayName("[성공] 본인이 등록한 학습자료를 모두 조회한다.")
+    void getMyPageDocument_본인이_등록한_학습자료를_모두_조회한다_성공() throws Exception {
+        final FileResponse fileResponse = new FileResponse(1L, "첨부파일명", "url");
+        final DocumentResponse document = DocumentResponse.builder()
+                .id(1L)
+                .title("학습자료 1")
+                .description("학습자료1 입니다")
+                .accessType(ALL)
+                .type(IMAGE)
+                .files(List.of(fileResponse))
+                .date(LocalDate.parse("2025-02-28"))
+                .uploaderName("홍길동")
+                .build();
+        final DocumentResponse otherDocument = DocumentResponse.builder()
+                .id(2L)
+                .title("학습자료 2")
+                .description("학습자료2 입니다")
+                .accessType(ALL)
+                .type(IMAGE)
+                .files(List.of(fileResponse))
+                .date(LocalDate.parse("2025-03-15"))
+                .uploaderName("홍길동")
+                .build();
+        final List<DocumentResponse> documents = List.of(document, otherDocument);
+
+        when(documentQueryService.getDocuments(any()))
+                .thenReturn(documents);
+
+        mockMvc.perform(get("/documents/members/me")
+                        .header(HttpHeaders.AUTHORIZATION, accessToken))
+                .andExpect(status().isOk())
+                .andDo(document("document-member-get-list"));
     }
 }
